+++ transportation-service/src/main/java/com/ftcs/transportation/trip_booking/controller/TripBookingsController.java
@@ -32,14 +32,14 @@ public class TripBookingsController {
     @PutMapping("/update/{bookingId}")
     @PreAuthorize("hasPermission(null, 'CUSTOMER')")
     public ApiResponse<String> updateTripBookings(@Valid @RequestBody TripBookingsRequestDTO requestDTO,
-                                                  @PathVariable("bookingId") Integer bookingId) {
+                                                  @PathVariable("bookingId") Long bookingId) {
         tripBookingsService.updateTripBookings(requestDTO, bookingId);
         return new ApiResponse<>("Trip booking updated successfully");
     }
 
     @PutMapping("/cancel/{bookingId}")
     @PreAuthorize("hasPermission(null, 'CUSTOMER')")
-    public ApiResponse<String> cancelTripBookings(@Valid @PathVariable("bookingId") Integer bookingId) {
+    public ApiResponse<String> cancelTripBookings(@Valid @PathVariable("bookingId") Long bookingId) {
         tripBookingsService.cancelTripBookings(bookingId);
         return new ApiResponse<>("Trip booking cancelled successfully");
     }
@@ -52,7 +52,7 @@ public class TripBookingsController {
     }
 
     @GetMapping("/{bookingId}")
-    public ApiResponse<TripBookings> getTripBookings(@PathVariable("bookingId") Integer bookingId) {
+    public ApiResponse<TripBookings> getTripBookings(@PathVariable("bookingId") Long bookingId) {
         TripBookings tripBookings = tripBookingsService.getTripBookings(bookingId);
         return new ApiResponse<>(tripBookings);
     }
@@ -68,7 +68,7 @@ public class TripBookingsController {
     @PreAuthorize("hasPermission(null, 'DRIVER')")
     public ApiResponse<String> updateStatusForDriver(@Valid @RequestBody UpdateStatusTripBookingsRequestDTO requestDTO,
                                                      @RequestAttribute("accountId") Integer accountId,
-                                                     @PathVariable("bookingId") Integer bookingId) {
+                                                     @PathVariable("bookingId") Long bookingId) {
         tripBookingsService.updateStatusForDriver(requestDTO, accountId, bookingId);
         return new ApiResponse<>("Driver status updated successfully");
     }
@@ -76,7 +76,7 @@ public class TripBookingsController {
     @PutMapping("/continueFindingDriver/{bookingId}")
     @PreAuthorize("hasPermission(null, 'CUSTOMER')")
     public ApiResponse<String> continueFindingDriver(@Valid @RequestBody UpdateStatusTripBookingsRequestDTO requestDTO,
-                                                     @PathVariable("bookingId") Integer bookingId) {
+                                                     @PathVariable("bookingId") Long bookingId) {
         tripBookingsService.continueFindingDriver(requestDTO, bookingId);
         return new ApiResponse<>("Continuing to find driver");
     }
@@ -85,7 +85,7 @@ public class TripBookingsController {
     @PreAuthorize("hasPermission(null, 'CUSTOMER') or hasPermission(null, 'DRIVER')")
     public ApiResponse<?> confirmCompleteDelivery(@Valid @RequestBody UpdateStatusTripBookingsRequestDTO requestDTO,
                                                   @RequestAttribute("role") String role,
-                                                  @PathVariable("bookingId") Integer bookingId) {
+                                                  @PathVariable("bookingId") Long bookingId) {
         tripBookingsService.confirmCompleteDelivery(requestDTO, role, bookingId);
         return new ApiResponse<>("Trip booking confirmed delivery successfully");
     }
