+++ transportation-service/src/main/java/com/ftcs/transportation/trip_matching/repository/DirectionsDataRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface DirectionsDataRepository extends JpaRepository<DirectionsData, Integer> {
+public interface DirectionsDataRepository extends JpaRepository<DirectionsData, Long> {
     DirectionsData findByStartLocationLatAndStartLocationLngAndEndLocationLatAndEndLocationLng(String startLocationLat, String startLocationLng, String endLocationLat, String endLocationLng);
 
     @Query(value = "EXEC dbo.GetUnmatchedSchedulesAndBookings", nativeQuery = true)
